Revert "all: smoother notification repository (fixes #7628) (#7604)"
This reverts commit 5c54c4c8cf5ab723416bd49001f985855e80dd27.
