+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/notification/AdapterNotification.kt
@@ -8,15 +8,17 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import java.util.regex.Pattern
-import kotlinx.coroutines.runBlocking
 import org.ole.planet.myplanet.R
 import org.ole.planet.myplanet.databinding.RowNotificationsBinding
+import org.ole.planet.myplanet.datamanager.DatabaseService
+import org.ole.planet.myplanet.model.RealmMyTeam
 import org.ole.planet.myplanet.model.RealmNotification
-import org.ole.planet.myplanet.repository.NotificationRepository
+import org.ole.planet.myplanet.model.RealmTeamTask
+import org.ole.planet.myplanet.model.RealmUserModel
 import org.ole.planet.myplanet.utilities.DiffUtils as DiffUtilExtensions
 
 class AdapterNotification(
-    private val notificationRepository: NotificationRepository,
+    private val databaseService: DatabaseService,
     notifications: List<RealmNotification>,
     private val onMarkAsReadClick: (String) -> Unit,
     private val onNotificationClick: (RealmNotification) -> Unit
@@ -98,26 +100,44 @@ class AdapterNotification(
                     } ?: notification.message
                 }
                 "join_request" -> {
-                    val metadata = runBlocking {
-                        notificationRepository.getJoinRequestMetadata(notification.relatedId)
+                    databaseService.withRealm { realm ->
+                        val joinRequest = realm.where(RealmMyTeam::class.java)
+                            .equalTo("_id", notification.relatedId)
+                            .equalTo("docType", "request")
+                            .findFirst()
+                        val team = joinRequest?.teamId?.let { tid ->
+                            realm.where(RealmMyTeam::class.java)
+                                .equalTo("_id", tid)
+                                .findFirst()
+                        }
+                        val requester = joinRequest?.userId?.let { uid ->
+                            realm.where(RealmUserModel::class.java)
+                                .equalTo("id", uid)
+                                .findFirst()
+                        }
+                        val requesterName = requester?.name ?: "Unknown User"
+                        val teamName = team?.name ?: "Unknown Team"
+                        "<b>${context.getString(R.string.join_request_prefix)}</b> " +
+                            context.getString(R.string.user_requested_to_join_team, requesterName, teamName)
                     }
-                    val requesterName = metadata?.requesterName ?: "Unknown User"
-                    val teamName = metadata?.teamName ?: "Unknown Team"
-                    "<b>${context.getString(R.string.join_request_prefix)}</b> " +
-                        context.getString(R.string.user_requested_to_join_team, requesterName, teamName)
                 }
                 else -> notification.message
             }
         }
 
         private fun formatTaskNotification(context: Context, taskTitle: String, dateValue: String): String {
-            val metadata = runBlocking { notificationRepository.getTaskNotificationMetadata(taskTitle) }
-            val message = context.getString(R.string.task_notification, taskTitle, dateValue)
-            val teamName = metadata?.teamName
-            return if (!teamName.isNullOrEmpty()) {
-                "<b>${teamName}</b>: $message"
-            } else {
-                message
+            return databaseService.withRealm { realm ->
+                val taskObj = realm.where(RealmTeamTask::class.java)
+                    .equalTo("title", taskTitle)
+                    .findFirst()
+                val team = realm.where(RealmMyTeam::class.java)
+                    .equalTo("_id", taskObj?.teamId)
+                    .findFirst()
+                if (team?.name != null) {
+                    "<b>${team.name}</b>: ${context.getString(R.string.task_notification, taskTitle, dateValue)}"
+                } else {
+                    context.getString(R.string.task_notification, taskTitle, dateValue)
+                }
             }
         }
     }
