+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/notification/NotificationsFragment.kt
@@ -88,7 +88,7 @@ class NotificationsFragment : Fragment() {
         }
 
         adapter = AdapterNotification(
-            notificationRepository,
+            databaseService,
             notifications,
             onMarkAsReadClick = { notificationId ->
                 markAsReadById(notificationId)
